+++ R/get_predictions_clmm.R
@@ -1,9 +1,7 @@
-#' @importFrom stats confint
-#' @importFrom insight find_random find_predictors print_color
 get_predictions_clmm <- function(model, terms, value_adjustment, condition, ci.lvl, linv, ...) {
 
   if (!requireNamespace("emmeans")) {
-    stop("Package `emmeans` required to compute marginal effects for clmm-models.", call. = FALSE)
+    stop("Package `emmeans` required to compute estimated marginal means for clmm-models.", call. = FALSE)
   }
 
   values.at <- .data_grid(