+++ R/getter.R
@@ -17,37 +17,38 @@
 #' @return The titles or labels as character string, or \code{NULL}, if variables
 #'         had no labels; \code{get_complete_df()} returns the input list \code{x}
 #'         as single data frame, where the grouping variable indicates the
-#'         marginal effects for each term.
+#'         predicted values for each term.
 #'
 #' @examples
-#' library(sjmisc)
-#' data(efc)
-#' efc$c172code <- to_factor(efc$c172code)
-#' fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
+#' if (require("sjmisc", quietly = TRUE) &&
+#'     require("ggplot2", quietly = TRUE) &&
+#'     require("effects", quietly = TRUE)) {
+#'   data(efc)
+#'   efc$c172code <- to_factor(efc$c172code)
+#'   fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
 #'
-#' mydf <- ggpredict(fit, terms = c("c12hour", "c161sex", "c172code"))
+#'   mydf <- ggpredict(fit, terms = c("c12hour", "c161sex", "c172code"))
 #'
-#' library(ggplot2)
-#' ggplot(mydf, aes(x = x, y = predicted, colour = group)) +
-#'   stat_smooth(method = "lm") +
-#'   facet_wrap(~facet, ncol = 2) +
-#'   labs(
-#'     x = get_x_title(mydf),
-#'     y = get_y_title(mydf),
-#'     colour = get_legend_title(mydf)
-#'   )
+#'   ggplot(mydf, aes(x = x, y = predicted, colour = group)) +
+#'     stat_smooth(method = "lm") +
+#'     facet_wrap(~facet, ncol = 2) +
+#'     labs(
+#'       x = get_x_title(mydf),
+#'       y = get_y_title(mydf),
+#'       colour = get_legend_title(mydf)
+#'     )
 #'
-#' # get marginal effects, a list of data frames (one data frame per term)
-#' eff <- ggeffect(fit)
-#' eff
-#' get_complete_df(eff)
+#'   # adjusted predictions, a list of data frames (one data frame per term)
+#'   eff <- ggeffect(fit)
+#'   eff
+#'   get_complete_df(eff)
 #'
-#' # get marginal effects for education only, and get x-axis-labels
-#' mydat <- eff[["c172code"]]
-#' ggplot(mydat, aes(x = x, y = predicted, group = group)) +
-#'   stat_summary(fun = sum, geom = "line") +
-#'   scale_x_discrete(labels = get_x_labels(mydat))
-#'
+#'   # adjusted predictions for education only, and get x-axis-labels
+#'   mydat <- eff[["c172code"]]
+#'   ggplot(mydat, aes(x = x, y = predicted, group = group)) +
+#'     stat_summary(fun = sum, geom = "line") +
+#'     scale_x_discrete(labels = get_x_labels(mydat))
+#' }
 #' @export
 get_title <- function(x, case = NULL) {
   if (.is_empty(x)) return(NULL)
@@ -126,11 +127,10 @@ get_x_labels <- function(x, case = NULL) {
 
 
 #' @rdname get_title
-#' @importFrom sjlabelled as_numeric
 #' @export
 get_complete_df <- function(x, case = NULL) {
   suppressWarnings(do.call(rbind, lapply(x, function(df) {
-    df$x <- sjlabelled::as_numeric(df$x)
+    df$x <- .factor_to_numeric(df$x)
     df
   })))
 }