+++ R/utils_model_function.R
@@ -1,4 +1,3 @@
-#' @importFrom insight model_info is_multivariate
 .get_model_function <- function(model) {
   # check class of fitted model
 
@@ -29,6 +28,7 @@ get_predict_function <- function(model) {
   if (inherits(model, c("wblm", "wbm"))) return("wbm")
   else if (inherits(model, "mclogit")) return("mclogit")
   else if (inherits(model, "averaging")) return("averaging")
+  else if (inherits(model, "orm")) return("orm")
   else if (inherits(model, "mlogit")) return("mlogit")
   else if (inherits(model, "glimML")) return("glimML")
   else if (inherits(model, "cgam")) return("cgam")